pyproject.toml
[tool.poetry]
name = "komtek"
version = "0.1.0"
description = ""
authors = ["vladimir <vladimir1012kov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.4"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
